==> 1. Data Cleaning.ipynb <==
import pandas as pd
import numpy as np

# Loading Data

brew = pd.read_csv('./data/breweryinfo.csv')
brew.head()

#### Let's remove the '%' and '/5' from brewery_pdev and brewery_score

brew['brewery_pdev'] = pd.to_numeric(brew['brewery_pdev'].str.replace('%', ''))

brew['brewery_score'] = pd.to_numeric(brew['brewery_score'].str.replace('/5', ''))

brew.head()

beer = pd.read_csv('./data/beerinfo.csv')
beer.head()

Reading the beer csv file prompted an error on two beers that were not parsed by the spider properly. The first one is [Weird Beard / Farmageddon Suspect Device](https://www.beeradvocate.com/beer/profile/31624/228703/) and the second one is [Interboro / Pipeworks - Mad Fat! Unicorn](https://www.beeradvocate.com/beer/profile/44293/272459/). I have no idea the scraper failed on those two, but we have to move on with our lives. We have to remove those entries. 

beer = beer[pd.to_numeric(beer['beer_number'], errors='coerce').notnull()]
beer['beer_number'] = pd.to_numeric(beer['beer_number'])

beer = beer[pd.to_numeric(beer['brewery_number'], errors='coerce').notnull()]
beer['brewery_number'] = pd.to_numeric(beer['brewery_number'])
beer.shape

comments = pd.read_csv('./data/comment.csv')

comments.head()

comments.shape

Similarly, we'd have to remove comment entries for those two beers

comments = comments[pd.to_numeric(comments['beer_number'], errors='coerce').notnull()]
comments['beer_number'] = pd.to_numeric(comments['beer_number'])

comments.shape

### There are some breweries with multiple locations

brew['brewery_name'].value_counts().head()

brew[brew['brewery_name'] == 'Granite City Food & Brewery'].head()

These appear to be Breweries that also own resturants. Luckily, it seems that their beer ratings are the same. We will ignore them for now, and will need additional work when we do analysis on them.

## Taking a look at comment data

comments['comment'].iloc[3]

#### There are evidently still some html tags left in the comments. We shell remove that

comments['comment'] = comments['comment'].str.replace('<[^>]*>',' ')
comments['comment'] = comments['comment'].str.replace('\\n',' ')
comments['comment'] = comments['comment'].str.replace('(\\xa0.*characters)',' ')

comments['comment'].iloc[3]

### Saving data as one SQL database

from sqlalchemy import create_engine
engine = create_engine('sqlite:///./data/beeradvocate.db')

brew.to_sql('breweries', con = engine)
beer.to_sql('beers', con = engine)
comments.to_sql('reviews', con = engine)

### Saving data as CSV files

brew.to_csv("./data/cleaned_brewerys.csv")
beer.to_csv("./data/cleaned_beers.csv")
comments.to_csv("./data/cleaned_reviews.csv")

## Subsetting data to Canadian only

can_brew = brew[brew['country'] == 'Canada']
can_brew.head()

can_beer = beer[beer['brewery_number'].isin(can_brew['brewery_number'])]
can_beer.head()

can_reviews = comments[comments['beer_number'].isin(can_beer['beer_number'])]
can_reviews.head()

### Save these files as csv for furthur analysis

can_brew.to_csv("./data/CanadianBreweries.csv")
can_beer.to_csv("./data/CanadianBeers.csv")
can_reviews.to_csv("./data/CanadianReviews.csv")
