==> 6. Topic Modelling.ipynb <==
import pandas as pd
brew = pd.read_csv("./data/CanadianBreweries.csv")
beer = pd.read_csv("./data/CanadianBeers.csv")
reviews = pd.read_csv("./data/CanadianReviews.csv")

#### Seperating the comment data

comments = reviews['comment']
comments.head()

#### Getting rid of NA

comments = comments.dropna()
comments.head()

#### Importing tools

import gensim
from gensim.utils import simple_preprocess
from gensim.parsing.preprocessing import STOPWORDS
from nltk.stem import WordNetLemmatizer, SnowballStemmer
from nltk.stem.porter import *
import numpy as np
import nltk
nltk.download('wordnet')

#### Word Processing

We need to process the data for topic modelling, which includes splitting sentences into words, removing punctuation, lowercasing the words, removing stopwords, lemmatizing words, removing stop word, and stemming the words

Lemmatizing means changing the word to first person form and present form.
Stemming means reducing words to their root form.

stemmer = SnowballStemmer('english')
def lemmatize_stemming(text):
    return stemmer.stem(WordNetLemmatizer().lemmatize(text, pos='v'))

def preprocess(text):
    result = []
    for token in gensim.utils.simple_preprocess(text):
        if token not in gensim.parsing.preprocessing.STOPWORDS and len(token) > 3:
            result.append(lemmatize_stemming(token))
    return result

Cheking the function

sample = comments.iloc[0]
sample

preprocess(sample)

#### Now is the time to process all comments. Warning: this step takes a while

processed_doc = comments.apply(preprocess)

processed_doc.head()

#### Build a dictionary from the corpus, and remove common words that are in over 50% of the comments, and remove those that appear less than 15 times

dictionary = gensim.corpora.Dictionary(processed_doc)

dictionary.filter_extremes(no_below=5, no_above=0.35, keep_n=100000)

#### Saving the corpus into a bag of words (BoW) format

bow_corpus = [dictionary.doc2bow(doc) for doc in processed_doc]

#### Running LDA using Bow with 8 threads

lda_model_20 = gensim.models.LdaMulticore(bow_corpus, num_topics=20, id2word=dictionary, passes=2, workers=7)
lda_model_30 = gensim.models.LdaMulticore(bow_corpus, num_topics=30, id2word=dictionary, passes=2, workers=7)

for idx, topic in lda_model_20.print_topics(-1):
    print('Topic: {} \nWords: {}'.format(idx, topic))

for idx, topic in lda_model_30.print_topics(-1):
    print('Topic: {} \nWords: {}'.format(idx, topic))

#### How many styles of beers are recorded in Canada

beer['style'].value_counts()

#### Using HDP to determin "optimal model numer"

from gensim.test.utils import common_corpus, common_dictionary
from gensim.models import HdpModel
hdp = HdpModel(bow_corpus, dictionary)

## Trying to seperate beer reviews by style

hdp.save('CanadianHDP')

hdp.print_topics(num_topics=50, num_words=10)

filtered_beer = beer[beer['style'] == 'Belgian Saison']

filtered_beer = filtered_beer.append(beer[beer['style'] == 'American IPA'])

filtered_beer = filtered_beer.append(beer[beer['style'] == 'Fruit and Field Beer'])

filtered_reviews = reviews[reviews['beer_number'].isin(filtered_beer['beer_number'])]

filtered_comments = filtered_reviews['comment'].dropna()

processed_doc = filtered_comments.apply(preprocess)

dictionary = gensim.corpora.Dictionary(processed_doc)

dictionary.filter_extremes(no_below=10, no_above=0.4, keep_n=100000)

bow_corpus = [dictionary.doc2bow(doc) for doc in processed_doc]

lda_model_5 = gensim.models.LdaMulticore(bow_corpus, num_topics=5, id2word=dictionary, passes=2, workers=7)

for idx, topic in lda_model_5.print_topics(-1):
    print('Topic: {} \nWords: {}'.format(idx, topic))
